Drop existing CalendarState.kt in favor of core interface
